reorganize relatedness files
