+++ src/Relatedness/RelatednessMoments.jl
@@ -1,4 +1,4 @@
-export QuellerGoodnight, Ritland, Lynch, LynchRitland, LynchLi, LiHorvitz, Moran, Blouin, Loiselle, Wang, relatedness
+export QuellerGoodnight, Ritland, Lynch, LynchRitland, LynchLi, LiHorvitz, Moran, Blouin, Loiselle, Wang
 
 """
     Blouin(ind1::GenoArray, ind2::GenoArray, locus_names::Vector{Symbol}; alleles::NamedTuple)
@@ -429,128 +429,4 @@ function Wang(ind1::T, ind2::T, locus_names::Vector{Symbol}, alleles::U; kwargs.
 
     r = (Φ/2.0 + Δ)
     return r
-end
-
-
-"""
-    relatedness(data::PopData; method::Function, iterations::Int64, interval::Tuple{Float64, Float64}, resample::String)
-    relatedness(data::PopData; method::Vector{Function}, iterations::Int64, interval::Tuple{Float64, Float64}, resample::String)
-    # to specify individuals for comparison
-    relatedness(data::PopData, sample_names::Vector{String}; method::Function, iterations::Int64, interval::Tuple{Float64, Float64}, resample::String)
-    relatedness(data::PopData, sample_names::Vector{String}; method::Vector{Function}, iterations::Int64, interval::Tuple{Float64, Float64}, resample::String)
-
-Return a dataframe of pairwise relatedness estimates for all or select pairs of samples in a `PopData` object. **Note:** samples must be diploid.
-
-### estimator methods
-There are several estimators available and are listed below. `relatedness` takes the
-function names as arguments (**case sensitive**), therefore do not use quotes or colons
-in specifying the methods. Methods can be supplied as a vector.
-- `Blouin`
-- `LiHorvitz`
-- `Loiselle`
-- `Lynch`
-- `LynchLi`
-- `LynchRitland`
-- `Moran`
-- `QuellerGoodnight`
-- `Ritland`
-- `Wang`
-
-For more information on a specific method, please see the respective docstring (e.g. `?Loiselle`).
-
-### bootstrap methods
-To calculate means, median, standard error, and confidence intervals using bootstrapping,
-set `iterations = n` where `n` is an integer greater than `0` (the default) corresponding to the number
-of bootstrap iterations you wish to perform for each pair. The default confidence interval is `(0.05, 0.95)` (i.e. 90%),
-however that can be changed by supplying a `Tuple` of `(low, high)` to the keyword `interval`.
-There are two available resampling methods, `"all"` (default) and `"nonmissing"`.
-- `"all"` : resamples all loci for a pair of individuals and then drops missing loci between them (recommended)
-    - speed: slower
-    - pro: better resampling variation
-    - con: by chance some iterations may have a lot of missing loci that have to be dropped
-- `"nonmissing"` : resamples only the shared non-missing loci between the pair
-    - speed: faster
-    - pro: every iteration guarantees the same number of loci compared between the pair
-    - con: too-tight confidence intervals due to less possible variation
-
-**Examples**
-```
-julia> cats = nancycats();
-
-julia> relatedness(cats, method = Ritland, iterations = 100);
-27966×8 DataFrame. Omitted printing of 2 columns
-│ Row   │ sample_1 │ sample_2 │ n_loci │ Ritland    │ Ritland_mean │ Ritland_median │
-│       │ String   │ String   │ Int64  │ Float64?   │ Float64?     │ Float64?       │
-├───────┼──────────┼──────────┼────────┼────────────┼──────────────┼────────────────┤
-│ 1     │ N215     │ N216     │ 8      │ 0.258824   │ 0.274537     │ 0.266457       │
-│ 2     │ N215     │ N217     │ 8      │ 0.193238   │ 0.191591     │ 0.180821       │
-│ 3     │ N215     │ N218     │ 8      │ 0.127497   │ 0.119988     │ 0.105559       │
-│ 4     │ N215     │ N219     │ 8      │ 0.0453471  │ 0.0558557    │ 0.0573132      │
-│ 5     │ N215     │ N220     │ 8      │ 0.108251   │ 0.12274      │ 0.110878       │
-│ 6     │ N215     │ N221     │ 8      │ 0.205139   │ 0.206509     │ 0.204635       │
-⋮
-│ 27961 │ N297     │ N281     │ 7      │ -0.0487076 │ -0.052506    │ -0.0549532     │
-│ 27962 │ N297     │ N289     │ 7      │ 0.154745   │ 0.170053     │ 0.169637       │
-│ 27963 │ N297     │ N290     │ 7      │ 0.189647   │ 0.19302      │ 0.189994       │
-│ 27964 │ N281     │ N289     │ 8      │ 0.0892068  │ 0.0914532    │ 0.0775897      │
-│ 27965 │ N281     │ N290     │ 7      │ 0.104614   │ 0.107821     │ 0.115087       │
-│ 27966 │ N289     │ N290     │ 7      │ 0.0511663  │ 0.0498539    │ 0.0547733      │
-
-julia> relatedness(cats, ["N7", "N111", "N115"], method = [Ritland, Wang]);
-3×5 DataFrame
-│ Row │ sample_1 │ sample_2 │ n_loci │ Ritland    │ Wang      │
-│     │ String   │ String   │ Int64  │ Float64?   │ Float64?  │
-├─────┼──────────┼──────────┼────────┼────────────┼───────────┤
-│ 1   │ N7       │ N111     │ 9      │ -0.129432  │ -0.395806 │
-│ 2   │ N7       │ N115     │ 9      │ -0.0183925 │ 0.0024775 │
-│ 3   │ N111     │ N115     │ 9      │ 0.0240152  │ 0.183966  │
-
-julia> relatedness(cats, ["N7", "N111", "N115"], method = [Loiselle, Moran], iterations = 100, interval = (0.025, 0.975));
-3×13 DataFrame. Omitted printing of 7 columns
-│ Row │ sample_1 │ sample_2 │ n_loci │ Loiselle   │ Loiselle_mean │ Loiselle_median │
-│     │ String   │ String   │ Int64  │ Float64?   │ Float64?      │ Float64?        │
-├─────┼──────────┼──────────┼────────┼────────────┼───────────────┼─────────────────┤
-│ 1   │ N7       │ N111     │ 9      │ -0.101618  │ 0.0141364     │ 0.00703954      │
-│ 2   │ N7       │ N115     │ 9      │ -0.0428898 │ 0.0743497     │ 0.0798708       │
-│ 3   │ N111     │ N115     │ 9      │ 0.13681    │ 0.266043      │ 0.257748        │
-```
-"""
-function relatedness(data::PopData, sample_names::Vector{String}; method::F, iterations::Int64 = 0, interval::Tuple{Float64, Float64} = (0.025, 0.975), resample::String = "all") where F
-    all(data.meta[data.meta.name .∈ Ref(sample_names), :ploidy] .== 2) == false && error("Relatedness analyses currently only support diploid samples")
-    errs = ""
-    all_samples = samples(data)
-    if sample_names != all_samples
-        for i in sample_names
-            if i ∉ all_samples
-                errs *= " $i,"
-            end
-        end
-        errs != "" && error("Samples not found in the PopData: " * errs)
-    end
-    if eltype(method) != Function
-        method = [method]
-    end
-    for i in Symbol.(method)
-        if i ∉ [:QuellerGoodnight, :Ritland, :Lynch, :LynchLi, :LynchRitland, :Wang, :Loiselle, :Blouin, :Moran, :LiHorvitz, :dyadicLikelihood]
-            errs *= "$i is not a valid method\n"
-        end
-    end
-    errs != "" && error(errs * "Methods are case-sensitive. Please see the docstring (?relatedness) for additional help.")
-    if iterations > 0
-        if resample == "all"
-            relatedness_boot_all(data, sample_names, method = method, iterations = iterations, interval = interval)
-        elseif resample == "nonmissing"
-            relatedness_boot_nonmissing(data, sample_names, method = method, iterations = iterations, interval = interval)
-        else
-            throw(ArgumentError("Please choose from resample methods \"all\" or \"nonmissing\""))
-        end
-    else
-        relatedness_no_boot(data, sample_names, method = method)
-    end
-end
-
-
-function relatedness(data::PopData; method::F, iterations::Int64 = 0, interval::Tuple{Float64, Float64} = (0.025, 0.975), resample::String = "all") where F
-    sample_names = samples(data) |> collect
-    relatedness(data, sample_names, method = method, iterations = iterations, interval = interval, resample = resample)
 end
